feat: extra docs for data blocks source + relations functionality
